requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django_db==0.0.7
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.11
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
